Replace innerHTML of parent element
In order to not actually touch the DOM until the very end it's doing some
kind of inefficient stuff, like calling document.getElementById in every
scroll listener, since nodes are getting created redundantly and immediately
destroyed the way it's implemented right now.

Also the scroll listeners. Should probably be one single listener that
checks all the questions instead of a separate listener for each one. But
it's so functional!